model-training/requirements.txt
# Core ML Libraries
tensorflow==2.15.0
keras==2.15.0
numpy==1.24.3
pandas==2.1.3

# Image Processing
pillow==10.1.0
opencv-python==4.8.1.78
scikit-image==0.22.0
albumentations==1.3.1

# Data Science
scikit-learn==1.3.2
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0

# Kaggle API
kaggle==1.5.16
opendatasets==0.1.22

# Utilities
tqdm==4.66.1
pyyaml==6.0.1
python-dotenv==1.0.0
loguru==0.7.2

# Jupyter
#jupyter==1.0.0
#ipykernel==6.27.1
#ipywidgets==8.1.1

# Model Analysis
tensorboard==2.15.1
keras-tuner==1.4.6

# Production
onnx==1.15.0
onnxruntime==1.16.3